Add safeguard for fast consecutive generate calls (#21)
If we call generate while there is already a generate call ongoing we should not generate immediately again which could block everything if called in short succession. For example when changing the random seed in the details panel.

With the new solution we set a flag that we want to regenerate and do this after the generate call has returned. We should also consider to create a "cancellation token" which could be used to stop the UnrealCallbacks from encoding all the geometry/materials which is unnecessary as we won't use the results anyways.