Move creating static meshes into tick function (#91)
Running it in a GameThread callback was not enough as for example during map change the callback could still be called even after the map has been unloaded which will result in invalid state. This way we only try to convert geometry if the tick function of the component is called, which also means we still have all valid state.