Fix GC issue with GCObject being constructed/destructed asynchronous
Remove GCObject from FVitruvioMesh because it is constructed in a background thread which can cause GC issues (if the GC runs at the same time the object is constructed). Instead we add it to a global list of used StaticMeshes in the VitruvioModule (and remove it once it is deconstructed). The VitruvioModule then handles the GC referenced objects.