Convert grayscale images to rgba...
This costs additional memory, but is the workaround with the least additional technical debt.
The reason for this change is that the sampler from a texture param node does not automatically update for transient textures in UE.
Note: Eventough the 16 bit format is called A16B16G16R16 the alpha channel is in the last memory location.