Serialize initial shape geometry and make initial shape SceneComponent truly transient
Before this change the IntialShape itself was responsible for saving and restoring its scene component (StaticMesh- or SplineComponent). This led to several issues where undo/redo operations would wrongly restore the scene component under certain circumstances. We now properly serialize the initial shape (with its geometry, which is also what is passed to PRT) instead of saving the scene component. We also make the VitruvioComponent responsible for creating a transient version of the scene component for visualization purpose on demand.