Fix float attribute editor crash
After pressing enter in a float editor the value is committed the model is regenerated and then the attribute editor is rebuilt. This causes a focus lost event on the old editor widget which also commits the value, but at this point the underlying property is not valid anymore which caused a segmentation fault.