Don't save generated static meshes
It does not make sense to save the StaticMesh InitialShapes as assets, since we already serialize the polygon of the InitialShape.

Saving them caused issues when the asset was not in the project anymore for packaged executables. We now just store a SoftReference to the asset to minimize creation of the StaticMesh if the Asset is still available.

Also we need to set the bFastBuild path for non editor builds (asserts otherwise).