Support builtin assets for replacements
This is useful if for example primitiveCubes are used as placeholders for replacements.
This means we now have to use the new meshId (which is the URI for filesystem assets or a new synthetic id constructed with the builtin asset name and the current leaf shape name) to identify instances.